+++ src/common/audit-builder.js
@@ -18,9 +18,11 @@ import {
   defaultMessageSender,
   defaultPersister,
   defaultUrlResolver,
+  defaultJobProvider,
   defaultPostProcessors,
   StepAudit,
   RunnerAudit,
+  AsyncJobRunner,
 } from './index.js';
 
 export class AuditBuilder {
@@ -31,6 +33,8 @@ export class AuditBuilder {
     this.persister = defaultPersister;
     this.messageSender = defaultMessageSender;
     this.postProcessors = defaultPostProcessors;
+    this.isAsyncJob = false;
+    this.jobProvider = defaultJobProvider;
     this.steps = {};
   }
 
@@ -65,6 +69,11 @@ export class AuditBuilder {
 
   withPostProcessors(postprocessors) {
     this.postProcessors = postprocessors;
+    return this;
+  }
+
+  withAsyncJob() {
+    this.isAsyncJob = true;
     return this;
   }
 
@@ -132,6 +141,19 @@ export class AuditBuilder {
           throw new Error(`Step ${stepName} must specify a destination as it is not the last step`);
         }
       });
+
+      if (this.isAsyncJob) {
+        return new AsyncJobRunner(
+          this.jobProvider,
+          this.siteProvider,
+          this.orgProvider,
+          this.urlResolver,
+          this.persister,
+          this.messageSender,
+          this.postProcessors,
+          this.steps,
+        );
+      }
 
       return new StepAudit(
         this.siteProvider,