+++ src/forms-opportunities/oppty-handlers/low-navigation-handler.js
@@ -12,6 +12,7 @@
 
 import { FORM_OPPORTUNITY_TYPES } from '../constants.js';
 import { filterForms, generateOpptyData } from '../utils.js';
+import { DATA_SOURCES } from '../../common/constants.js';
 
 const formPathSegments = ['contact', 'newsletter', 'sign', 'enrol', 'subscribe', 'register', 'join', 'apply', 'quote', 'buy', 'trial', 'demo', 'offer'];
 /**
@@ -22,7 +23,9 @@ const formPathSegments = ['contact', 'newsletter', 'sign', 'enrol', 'subscribe',
  */
 // eslint-disable-next-line max-len
 export default async function createLowNavigationOpportunities(auditUrl, auditDataObject, scrapedData, context, excludeForms = new Set()) {
-  const { dataAccess, log } = context;
+  const {
+    dataAccess, log, sqs, site, env,
+  } = context;
   const { Opportunity } = dataAccess;
 
   // eslint-disable-next-line no-param-reassign
@@ -74,6 +77,7 @@ export default async function createLowNavigationOpportunities(auditUrl, auditDa
         tags: ['Forms Conversion'],
         data: {
           ...opptyData,
+          dataSources: [DATA_SOURCES.RUM, DATA_SOURCES.PAGE],
         },
         guidance: {
           recommendations: [
@@ -101,6 +105,25 @@ export default async function createLowNavigationOpportunities(auditUrl, auditDa
         // eslint-disable-next-line no-await-in-loop
         await highPageViewsLowFormNavOppty.save();
       }
+
+      log.info('sending message to mystique for high-page-views-low-form-nav');
+      const mystiqueMessage = {
+        type: 'guidance:high-page-views-low-form-nav',
+        siteId: auditData.siteId,
+        auditId: auditData.auditId,
+        deliveryType: site.getDeliveryType(),
+        time: new Date().toISOString(),
+        data: {
+          url: opportunityData.data.form,
+          cr: opportunityData.data.trackedFormKPIValue,
+          screenshot: opportunityData.data.screenshot,
+          cta_source: opportunityData.data.formNavigation.source,
+        },
+      };
+
+      // eslint-disable-next-line no-await-in-loop
+      await sqs.sendMessage(env.QUEUE_SPACECAT_TO_MYSTIQUE, mystiqueMessage);
+      log.info(`forms opportunity high page views low form nav sent to mystique: ${JSON.stringify(mystiqueMessage)}`);
     }
   } catch (e) {
     log.error(`Creating Forms opportunity for high page views low form nav for siteId ${auditData.siteId} failed with error: ${e.message}`, e);