+++ test/fixtures/forms/high-form-views-low-conversions.js
@@ -133,6 +133,62 @@ const testData = {
       ],
     },
   },
+  opptyAuditDataWithIframe: {
+    type: 'high-page-views-low-form-nav',
+    siteId: 'site-id',
+    auditId: 'audit-id',
+    auditResult: {
+      formVitals: [
+        {
+          url: 'https://www.iframe-example.com/test/getting-iframe-example/guide/newsletter',
+          formsubmit: {
+          },
+          formview: { 'mobile:ios': 300 },
+          formengagement: {
+            'mobile:ios': 300,
+          },
+          pageview: {
+            'desktop:windows': 4670,
+            'mobile:ios': 4000,
+          },
+          trafficacquisition: {
+            paid: 4670,
+            maxTimeDelta: 3060,
+            total: 8670,
+            earned: 2000,
+            sources: [],
+            owned: 2000,
+          },
+          iframeSrc: 'https://www.iframe-example.com/content/iframe-example/en-us/test/getting-iframe-example/guide/begin/jcr:content/contentpar/columns/0/aemform.iframe.en.html',
+          forminternalnavigation: [
+            {
+              url: 'https://www.suriframe-example.com/newsletter',
+              pageview: {
+                'desktop:windows:blink': 54000,
+                'mobile:android:blink': 26000,
+                'mobile:ios:webkit': 24000,
+                'desktop:mac:webkit': 2000,
+                'desktop:chromeos:blink': 900,
+                'desktop:mac:blink': 900,
+                'desktop:linux:gecko': 200,
+                'mobile:ipados:webkit': 100,
+                'mobile:android:gecko': 100,
+                'desktop:linux:blink': 100,
+                'desktop:windows:gecko': 100,
+              },
+              CTAs: [
+                {
+                  source: '#teaser-related02 .cmp-teaser__action-link',
+                  clicks: 800,
+                },
+              ],
+              totalClicksOnPage: 7200,
+            },
+          ],
+        },
+      ],
+    },
+  },
   oppty2AuditData: {
     type: 'high-page-views-low-form-nav',
     siteId: 'site-id',
@@ -496,7 +552,7 @@ const testData = {
           formsubmit: {
             'desktop:windows': 100,
           },
-          formview: { 'mobile:ios': 3000 },
+          formview: { 'mobile:ios': 3200 },
           formengagement: {
             'mobile:ios': 300,
           },
@@ -625,7 +681,7 @@ const testData = {
           },
           formview: {
             'desktop:windows': 2000,
-            'mobile:ios': 1000,
+            'mobile:ios': 1200,
           },
           formengagement: {
             'desktop:windows': 700,
@@ -741,20 +797,20 @@ const testData = {
     ],
     data: {
       trackedFormKPIName: 'Conversion Rate',
-      trackedFormKPIValue: 0.03333333333333333,
+      trackedFormKPIValue: 0.03125,
       metrics: [
         {
           type: 'conversionRate',
           device: '*',
           value: {
-            page: 0.03333333333333333,
+            page: 0.03125,
           },
         },
         {
           type: 'bounceRate',
           device: '*',
           value: {
-            page: 0.9,
+            page: 0.90625,
           },
         },
         {
@@ -796,7 +852,7 @@ const testData = {
           type: 'bounceRate',
           device: 'mobile',
           value: {
-            page: 0.9,
+            page: 0.90625,
           },
         },
         {
@@ -808,22 +864,30 @@ const testData = {
         },
         {
           type: 'traffic',
-          device: '*',
+          device: 'desktop',
           value: {
-            total: null,
-            paid: null,
-            earned: null,
-            owned: null,
+            page: 4670,
+          },
+        },
+        {
+          type: 'traffic',
+          device: 'mobile',
+          value: {
+            page: 4000,
           },
         },
       ],
       form: 'https://www.surest.com/newsletter',
-      screenshot: '',
-      formViews: 3000,
-      pageViews: 8670,
       formsource: '',
+      formViews: 3200,
+      pageViews: 8670,
+      screenshot: '',
       samples: 8670,
       scrapedStatus: false,
+      dataSources: [
+        'RUM',
+        'Page',
+      ],
     },
     guidance: {},
   },
@@ -848,6 +912,7 @@ const testData = {
       formsource: '',
       samples: 6690,
       scrapedStatus: false,
+      dataSources: ['RUM', 'Page'],
       metrics: [
         {
           type: 'conversionRate',
@@ -914,12 +979,16 @@ const testData = {
         },
         {
           type: 'traffic',
-          device: '*',
+          device: 'desktop',
           value: {
-            total: null,
-            paid: null,
-            earned: null,
-            owned: null,
+            page: 5690,
+          },
+        },
+        {
+          type: 'traffic',
+          device: 'mobile',
+          value: {
+            page: 1000,
           },
         },
       ],
@@ -947,6 +1016,7 @@ const testData = {
       formsource: '',
       samples: 6690,
       scrapedStatus: false,
+      dataSources: ['RUM', 'Page'],
       metrics: [
         {
           type: 'conversionRate',
@@ -1013,12 +1083,16 @@ const testData = {
         },
         {
           type: 'traffic',
-          device: '*',
+          device: 'desktop',
           value: {
-            total: null,
-            paid: null,
-            earned: null,
-            owned: null,
+            page: 5690,
+          },
+        },
+        {
+          type: 'traffic',
+          device: 'mobile',
+          value: {
+            page: 1000,
           },
         },
       ],
@@ -1055,6 +1129,7 @@ const testData = {
       formsource: '',
       samples: 6690,
       scrapedStatus: true,
+      dataSources: ['RUM', 'Page'],
       metrics: [
         {
           type: 'conversionRate',
@@ -1121,12 +1196,16 @@ const testData = {
         },
         {
           type: 'traffic',
-          device: '*',
+          device: 'desktop',
           value: {
-            total: null,
-            paid: null,
-            earned: null,
-            owned: null,
+            page: 5690,
+          },
+        },
+        {
+          type: 'traffic',
+          device: 'mobile',
+          value: {
+            page: 1000,
           },
         },
       ],
@@ -1155,20 +1234,20 @@ const testData = {
     ],
     data: {
       trackedFormKPIName: 'Conversion Rate',
-      trackedFormKPIValue: 0.06666666666666667,
+      trackedFormKPIValue: 0.0625,
       metrics: [
         {
           type: 'conversionRate',
           device: '*',
           value: {
-            page: 0.06666666666666667,
+            page: 0.0625,
           },
         },
         {
           type: 'bounceRate',
           device: '*',
           value: {
-            page: 0.6666666666666667,
+            page: 0.6875,
           },
         },
         {
@@ -1203,14 +1282,14 @@ const testData = {
           type: 'conversionRate',
           device: 'mobile',
           value: {
-            page: 0.1,
+            page: 0.08333333333333333,
           },
         },
         {
           type: 'bounceRate',
           device: 'mobile',
           value: {
-            page: 0.7,
+            page: 0.75,
           },
         },
         {
@@ -1222,22 +1301,30 @@ const testData = {
         },
         {
           type: 'traffic',
-          device: '*',
+          device: 'desktop',
           value: {
-            total: 211400,
-            paid: 192200,
-            earned: 4400,
-            owned: 14800,
+            page: 5690,
+          },
+        },
+        {
+          type: 'traffic',
+          device: 'mobile',
+          value: {
+            page: 2000,
           },
         },
       ],
       form: 'https://www.surest.com/contact-us',
-      formViews: 3000,
-      pageViews: 7690,
       formsource: '',
+      formViews: 3200,
+      pageViews: 7690,
       screenshot: '',
       samples: 7690,
       scrapedStatus: false,
+      dataSources: [
+        'RUM',
+        'Page',
+      ],
     },
     guidance: {},
   },